requirements.txt
pygame
anthropic
python-dotenv
